+++ src/AbatabLieutenant/Program.cs
@@ -1,11 +1,11 @@
 ﻿/*************************************************************************
- * Abatab Lieutenant v4.0.0+230308.0939
+ * Abatab Lieutenant v4.1.0+230516
  * A command line deployment utility for Abatab.
  * https://github.com/spectrum-health-systems/AbatabLieutenant
  ************************************************************************/
 
 // AbatabLieutenant.Program.cs
-// b230308.0939
+// b230516.0955
 // (c) A Pretty Cool Program
 
 namespace AbatabLieutenant
@@ -19,13 +19,16 @@ namespace AbatabLieutenant
         {
             Console.Clear();
 
-            var ltSession = LtSession.LoadLocalSettings("AbatabLieutenant.json");
+            var ltSession = LtSession.LoadLocalSettings();
 
             if (commandArguments.Length > 0 && ltSession.ValidBranches.Contains(commandArguments[0]))
             {
                 LtSession.CreateRuntimeSettings(ltSession, commandArguments);
+
                 Utilities.VerifyFramework(ltSession.AbatabDataFolders, ltSession.LogPath);
+
                 Deploy.WebService(ltSession);
+
                 Utilities.WriteLog($"{Environment.NewLine}Deployment complete!", ltSession.LogPath);
             }
             else